requirements.txt
django==4.1.5
python-dotenv==0.21.0
psycopg2-binary==2.9.5
django-phonenumber-field==7.0.2
phonenumbers==8.13.4
django-djeddit==0.3.4
pipenv==2020.11.15
djangorestframework==3.12.2
djangorestframework-simplejwt
django-cors-headers==3.7.0
django-import-export==3.0.2
whitenoise==6.2.0
gunicorn==20.1.0
django-googledrive-storage==1.6.0
